requirements.txt
flask
flask-cors
pymongo
python-dotenv
flask-pymongo
bcrypt
pyjwt
scikit-learn
joblib
